refactor: moves stream logic outside server.c
- moves stream interface outside of server.c, letting it be configured
  by the library client and allowing for bindings of cgo to be injected
  there
