+++ pkg/quic/lsquic/CMakeLists.txt
@@ -23,12 +23,13 @@ IF(BUILD_LSQUIC)
     INCLUDE_DIRECTORIES(lsquic/src/liblsquic)
     INCLUDE_DIRECTORIES(include)
     ADD_LIBRARY(lsquic_adapter SHARED ./src/logger.c ./src/cert.c ./src/server.c ./src/keylog.c ./src/ancillary.c)
-    TARGET_LINK_LIBRARIES(lsquic_adapter PRIVATE
+    SET(LIBS
+        lsquic_adapter
         ${CMAKE_SOURCE_DIR}/lsquic/src/liblsquic/liblsquic.so
         ${CMAKE_SOURCE_DIR}/boringssl/crypto/libcrypto.so
         ${CMAKE_SOURCE_DIR}/boringssl/ssl/libssl.so z ev m)
     add_executable(echo_test.out cmd/echo_test.c)
-    target_link_libraries(echo_test.out lsquic_adapter)
+    target_link_libraries(echo_test.out ${LIBS})
 ENDIF()
 
 MESSAGE(STATUS "Compiler flags: ${CMAKE_C_FLAGS}")
